Add cleanup steps for stateful set.
This commit adds a small cleanup step for stateful set chapter.
It was easier to just delete the app via the yaml spec that was
applied earlier.

I also added a step to delete the persistent volume claim that was
created as a part of the yaml spec of `mehdb` app.